Move awair coordinators to their file
