+++ homeassistant/components/awair/sensor.py
@@ -31,7 +31,6 @@ from homeassistant.helpers.device_registry import DeviceInfo
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 from homeassistant.helpers.update_coordinator import CoordinatorEntity
 
-from . import AwairDataUpdateCoordinator, AwairResult
 from .const import (
     API_CO2,
     API_DUST,
@@ -46,6 +45,7 @@ from .const import (
     ATTRIBUTION,
     DOMAIN,
 )
+from .coordinator import AwairDataUpdateCoordinator, AwairResult
 
 DUST_ALIASES = [API_PM25, API_PM10]
